backfill/requirements.txt
solace-pubsubplus>=1.5.0
psycopg2-binary>=2.9.0
requests>=2.28.0
python-dotenv>=1.0.0
